add a monitoring common for alertmanagerconfig secret and create secrets in AMO and OO
