+++ pkg/products/codeready/reconciler.go
@@ -3,6 +3,7 @@ package codeready
 import (
 	"context"
 	"fmt"
+	"github.com/integr8ly/integreatly-operator/pkg/products/monitoringcommon"
 
 	l "github.com/integr8ly/integreatly-operator/pkg/resources/logger"
 	"github.com/integr8ly/integreatly-operator/pkg/resources/quota"
@@ -287,7 +288,7 @@ func (r *Reconciler) createResource(ctx context.Context, resourceName string, se
 	r.extraParams["MonitoringKey"] = r.Config.GetLabelSelector()
 	r.extraParams["Namespace"] = r.Config.GetNamespace()
 
-	templateHelper := monitoring.NewTemplateHelper(r.extraParams)
+	templateHelper := monitoringcommon.NewTemplateHelper(r.extraParams)
 	resource, err := templateHelper.CreateResource(resourceName)
 	if err != nil {
 		return nil, fmt.Errorf("createResource failed: %w", err)
